Fixed issue #3
Some entries didn't show up in the search results because the
computation of rows was made on a smaller data size, because this
data was chopped whenever the QByteArray response contained a null
character (\0).

As we can never be entirely sure of the content of the database,
we make the code more robust by removing all null characters from
the received QByteArray.