Fixed bug in search
When deleting a word online, a refresh was called immediately after the
delete request. So the reply of that second request could be received
before the reply of the first one, which led to showing the state of the
database before the deletion.

HTTP requests are now sequentially made: it is only when the reply to
the delete request is received that we call the refresh function.