Fixed bug in selected tags
When editing a word having no tag, the tag id list string is an empty
string. From that string the generated integer tag id list was [0]
instead of []. So then the database manager tried to assign 0 as a tag
id for the word and triggered a foreign key error.

Also swapped the tag id and status widgets.