Back button handling in search view
When TestView's back button is clicked, the action is first passed over
to the SearchView (if it exists) which will either remove its EditView
(if it exists) or leave the go back action for its parent to handle.

The net effect is that clicking on the back button from a SearchView's
EditView will go back to the SearchView instead of going back to the
TestView.